fix(schedules): restore pre-refactor parity flagged by adversarial review
- Retry defaults are resolved but no longer validated in the Scheduler
  constructor: a historically tolerated invalid static retry config must
  not start throwing in the Agent constructor and brick every entry point
  of the Durable Object. Invalid defaults surface per execution as
  schedule:error, as before; per-schedule retry overrides stay validated.
- One-shot idempotent dedup accepts any truthy value again (historic
  behavior), not only literal true.
- Scheduler storage failures throw the exported SqlError again; the class
  moved to sql-error.ts so agents/schedules can share it without a cycle.
- The startup schedule() warning message now says 'during startup' (the
  window deliberately covers all startup hooks, not just onStart), and the
  convention-based underscore-callback exemption is gone — the one internal
  startup caller passes an explicit idempotent choice instead.
- The capability test's cron advance assertion tolerates a minute-boundary
  landing on the current second; stale 'Lifecycle controller' wording fixed
  in docs.
